Restructuring Scancode-Toolkit Documentation According to 4 functions Rule
This commit completely changes the documentation structure, dividing it into sub-folder adhering to the 4 functions - tutorials, how-to
guides, referance and explanation. Also adds the Getting Started, Contributing and Miscellaneous sections. This commit contains only the index.rst
files, and later commits will add modular changes in this restructuring. This is very close to the structure that will remain at last, but a lot
of content will be added subsequently in the files, and different branches for different versions.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>