Adds New Tutorials and Updates other for Version 3.1.1
This commit adds to the Scancode Version 3.1.1 doc branch -> 2 main sample tutorials, How to Run a Scan, and
How to Visualize Scan Results. It also makes changes to How to Format Scan output and How to Set what will be
Detected in a Scan to be compatible with Scancode version 3.1.1, and adds images/gifs (May be updated later).
Also adds minor links/referances.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>